+++ upload/folderUploadJob.go
@@ -8,16 +8,14 @@ import (
 
 	gphotos "github.com/gphotosuploader/google-photos-api-client-go/lib-gphotos"
 
-	"github.com/gphotosuploader/gphotos-uploader-cli/datastore/completeduploads"
-	"github.com/gphotosuploader/gphotos-uploader-cli/datastore/uploadurls"
+	"github.com/gphotosuploader/gphotos-uploader-cli/app"
 	"github.com/gphotosuploader/gphotos-uploader-cli/utils/filesystem"
 )
 
 // Job represents a job to upload all photos from the specified folder
 type Job struct {
-	client            *gphotos.Client
-	trackingService   *completeduploads.Service
-	uploadURLsService *uploadurls.Service
+	gPhotos     *gphotos.Client
+	fileTracker app.FileTracker
 
 	sourceFolder string
 	options      *JobOptions
@@ -44,11 +42,10 @@ func NewJobOptions(createAlbum bool, deleteAfterUpload bool, uploadVideos bool, 
 }
 
 // NewFolderUploadJob creates a job based on the submitted data
-func NewFolderUploadJob(client *gphotos.Client, trackingService *completeduploads.Service, uploadURLsService *uploadurls.Service, fp string, opt *JobOptions) *Job {
+func NewFolderUploadJob(gPhotos *gphotos.Client, fileTracker app.FileTracker, fp string, opt *JobOptions) *Job {
 	return &Job{
-		trackingService:   trackingService,
-		uploadURLsService: uploadURLsService,
-		client:            client,
+		fileTracker: fileTracker,
+		gPhotos:     gPhotos,
 
 		sourceFolder: fp,
 		options:      opt,
@@ -68,7 +65,6 @@ func (job *Job) ScanFolder(uploadChan chan<- *Item) error {
 	// dirs are walked depth-first.   These vars hold the active album
 	// default empty album for makeAlbums.enabled = false
 	errW := filepath.Walk(job.sourceFolder, func(fp string, fi os.FileInfo, errP error) error {
-		// log.Printf("ScanFolder.Walk: %v, fi: %v, err: %v\n", fp, fi, err)
 		if err != nil {
 			_, _ = fmt.Fprintf(os.Stderr, "error for %v: %v\n", fp, err)
 			return nil
@@ -92,7 +88,7 @@ func (job *Job) ScanFolder(uploadChan chan<- *Item) error {
 		}
 
 		// check completed uploads db for previous uploads
-		isAlreadyUploaded, err := job.trackingService.IsAlreadyUploaded(fp)
+		isAlreadyUploaded, err := job.fileTracker.IsAlreadyUploaded(fp)
 		if err != nil {
 			log.Println(err)
 		} else if isAlreadyUploaded {
@@ -104,12 +100,12 @@ func (job *Job) ScanFolder(uploadChan chan<- *Item) error {
 		var albumID string
 		if job.options.createAlbum {
 			name := filepath.Base(filepath.Dir(fp))
-			albumID = getGooglePhotosAlbumID(name, job.client)
+			albumID = getGooglePhotosAlbumID(name, job.gPhotos)
 		}
 
 		// set file upload options depending on folder upload options
 		var uploadItem = &Item{
-			client:          job.client,
+			gPhotos:         job.gPhotos,
 			path:            fp,
 			album:           albumID,
 			deleteOnSuccess: job.options.deleteAfterUpload,
@@ -127,8 +123,8 @@ func (job *Job) ScanFolder(uploadChan chan<- *Item) error {
 	return nil
 }
 
-// getGooglePhotosAlbumID return the Id of an album with the specified name.
-// If the album doesn't exist, return an empty string.
+// getGooglePhotosAlbumID returns the ID of an album with the specified name.
+// If the album didn't exist, it's created thanks to GetOrCreateAlbumByName().
 func getGooglePhotosAlbumID(name string, c *gphotos.Client) string {
 	if name == "" {
 		return ""